# improve performance issue of `@nospecialize`-d keyword func call