@aviatesk 2022-10-05
@nanosoldier `runbenchmarks(!"scalar", vs=":master")`