@nanosoldier 2022-10-05
[Your benchmark job](https://github.com/JuliaLang/julia/pull/47059#issuecomment-1268408840) has completed - possible performance regressions were detected. A full report can be found [here](https://github.com/JuliaCI/NanosoldierReports/blob/master/benchmark/by_hash/ddec630_vs_fe7f673/report.md).