@Keno 2022-10-05
> %2 = NamedTuple{(:a, :b, :c)(%1)::NamedTuple{(:a, :b, :c), _A} where _A<:Tuple{Any, Any, Any}

Why doesn't this get inlined into `splatnew`? Are we just missing a case for it in the inlining cost model? We're also probably missing SROA support for `splatnew`, but I'd rather fix that than special-casing `NamedTuple` here.