@aviatesk 2022-10-07
> Why doesn't this get inlined into `splatnew`? Are we just missing a case for it in the inlining cost model?

This call is from 
https://github.com/JuliaLang/julia/blob/deb87b9ba9a939ff8a30551e0c9a3f20777ea551/base/boot.jl#L618
and the redirected constructor call is union split to:
- https://github.com/JuliaLang/julia/blob/deb87b9ba9a939ff8a30551e0c9a3f20777ea551/base/boot.jl#L622-L623
- https://github.com/JuliaLang/julia/blob/deb87b9ba9a939ff8a30551e0c9a3f20777ea551/base/namedtuple.jl#L90-L97

And the latter contains dynamic dispatch since `args` is abstract tuple type (and somehow the former is also judged as non-inlineable somehow).

---

@nanosoldier `runbenchmarks("inference", vs=":master")`