@aviatesk 2022-10-07
> Can we use the regular error check pattern to switch that around?

Well, the problematic dynamic dispatch is not that error path but `T(args)`, which is union-split to https://github.com/JuliaLang/julia/blob/deb87b9ba9a939ff8a30551e0c9a3f20777ea551/base/essentials.jl#L411-L420
and the latter split is confused for abstract tuple input.