@Keno 2022-10-07
Ah, ok, so the issue is that we have:
```
NamedTuple{names}(args::Tuple) where {names} = NamedTuple{names,typeof(args)}(args)
```
but inference looses the type constraint that the second type parameter is typeequal to `typeof(args)`, so it tries to unionsplit it into that mess. How about the following change then:
```
diff --git a/base/boot.jl b/base/boot.jl
index 5f3b99df1c..4e02725fc3 100644
--- a/base/boot.jl
+++ b/base/boot.jl
@@ -615,7 +615,8 @@ end
 
 NamedTuple() = NamedTuple{(),Tuple{}}(())
 
-NamedTuple{names}(args::Tuple) where {names} = NamedTuple{names,typeof(args)}(args)
+eval(Core, :(NamedTuple{names}(args::Tuple) =
+             $(Expr(:splatnew, :(NamedTuple{names,typeof(args)}), :args))))
 
 using .Intrinsics: sle_int, add_int
```
That should also save us some inference time by not having to infer through the useless unionsplit.