@aviatesk 2022-10-07
That sounds quite simple and better! I will work on implementing SROA for `:splatnew`.