@nanosoldier 2022-10-07
[Your benchmark job](https://github.com/JuliaLang/julia/pull/47059#issuecomment-1271318059) has completed - no performance regressions were detected. A full report can be found [here](https://github.com/JuliaCI/NanosoldierReports/blob/master/benchmark/by_hash/e094dd1_vs_d498d36/report.md).