@aviatesk 2022-10-08
Okay, this PR should be ready.
I found we actually have `inline_splatnew!` pass within the inlining that transforms `:splatnew` to `:new` expression. And the root problem turned out to be that our `nfields_tfunc` isn't accurate enough to handle abstract `NamedTuple` whose names are nevertheless fully known and thus the number of their fields could be known.
This PR improves that accuracy, and now our SROA pass optimizes away more `NamedTuple`s constructed for keyword argument handling.