@nanosoldier 2022-10-08
[Your benchmark job](https://github.com/JuliaLang/julia/pull/47059#issuecomment-1272194432) has completed - no performance regressions were detected. A full report can be found [here](https://github.com/JuliaCI/NanosoldierReports/blob/master/benchmark/by_hash/b3026b5_vs_4c0f8de/report.md).