@aviatesk 2022-10-08
The benchmark results look promising. Going to merge once confirm successful CI.