@aviatesk 2022-10-08
@nanosoldier `runbenchmarks(!"scalar", vs=":master")`