@nanosoldier 2022-10-08
[Your benchmark job](https://github.com/JuliaLang/julia/pull/47059#issuecomment-1272217976) has completed - possible performance regressions were detected. A full report can be found [here](https://github.com/JuliaCI/NanosoldierReports/blob/master/benchmark/by_hash/7c4d294_vs_df5b081/report.md).